Crate cu29_unifiedlog

Re-exports

pub use crate::memmap::MmapUnifiedLogger as UnifiedLogger;

pub use crate::memmap::MmapUnifiedLoggerBuilder as UnifiedLoggerBuilder;

pub use crate::memmap::MmapUnifiedLoggerRead as UnifiedLoggerRead;

pub use crate::memmap::MmapUnifiedLoggerWrite as UnifiedLoggerWrite;

pub use crate::memmap::UnifiedLoggerIOReader;

Modules

memmap

This is the memory map file implementation for the unified logger for Copper. It is std only.

Structs

EndOfLogMarker

Payload stored in the end-of-log section to signal whether the log was cleanly closed.

LogStream

A wrapper around the unifiedlogger that implements the Write trait.

MainHeader

The main file header of the datalogger.

SectionHandle

A SectionHandle is a handle to a section in the datalogger. It allows tracking the lifecycle of the section.

SectionHeader

Each concurrent sublogger is tracked through a section header. They form a linked list of sections. The entry type is used to identify the type of data in the section.

UnifiedLogStatus

Basic statistics for the unified logger. Note: the total_allocated_space might grow for the std implementation

Enums

AllocatedSection

Constants

MAIN_MAGIC

ID to spot the beginning of a Copper Log

SECTION_HEADER_COMPACT_SIZE

SECTION_MAGIC

ID to spot a section of Copper Log

Traits

SectionStorage

A Storage is an append-only structure that can update a header section.

UnifiedLogRead

Read back a unified log linearly.

UnifiedLogWrite

The writing interface to the unified logger. Writing is “almost” linear as various streams can allocate sections and track them until they drop them.

Functions

stream_write

Create a new stream to write to the unifiedlogger.