Crate cu29_unifiedlog

Structs

SectionHandle

A SectionHandle is a handle to a section in the datalogger. It allows to track the lifecycle of a section of the datalogger.

SectionHeader

Each concurrent sublogger is tracked through a section header. They form a linked list of sections. The entry type is used to identify the type of data in the section.

UnifiedLoggerBuilder

Use this builder to create a new DataLogger.

UnifiedLoggerIOReader

This a convenience wrapper around the UnifiedLoggerRead to implement the Read trait.

UnifiedLoggerRead

A read side of the datalogger.

UnifiedLoggerWrite

A write side of the datalogger.

Enums

AllocatedSection

UnifiedLogger

Holder of the read or write side of the datalogger.

Functions

stream_write

Create a new stream to write to the unifiedlogger.